implement triangle import kind 2
while trying to get this to work, I made some major changes.

We can now import the triangles with kind 2, the simplest of the packing
formats.

I need to clean up the result, I initially made some incorrect
assumptions (namely that the vertex buffer is gap-free) that lead to
sub-optimal code structure.