force balanced tree
By fixing the pivot after half the elements, we can minimise the tree
depth. But this may lead to large nodes with lots of overlap.

Of the six possible partitions, we choose the one with the least
overlap.

I don't understand why this generates a tree of depth 15 (32K elements)
for 9.6K triangles. Even if each leaf was just one triangle, that would
be overkill.