Add partial shadowmap import
For now, it only generates an image, but doesn't place it in the track.
I've tried aligning it manually for canal1a, the result is not something
I would know how to derive from the collision data. Maybe as far as the
buggy car brightness on re-exports goes, this is a red herring?